ユーザ選択UIを<UserSelectList>とし、リッチにする
