+++ frontend/src/features/Chat/Chat.tsx
@@ -56,7 +56,10 @@ export const Chat = (props: { mySocket: ReturnType<typeof io> }) => {
             <div className="flex w-full shrink grow flex-col overflow-hidden border-2 border-solid border-white">
               <Switcher />
               <div className="shrink-0 grow-0 p-2">
-                <FTButton className="w-full" onClick={openModal}>
+                <FTButton
+                  className="flex w-full flex-row justify-center"
+                  onClick={openModal}
+                >
                   <InlineIcon i={<Icons.Add />} />
                   新規作成
                 </FTButton>
