+++ frontend/src/features/Chat/RoomSetting.tsx
@@ -31,14 +31,14 @@ const RoomTypeListBox = ({ selected, setSelected }: RoomTypeListProps) => {
           setSelected(next.roomType)
         }
       >
-        <Listbox.Button className="w-[9em] border-2 pl-2 pr-4 text-center">
+        <Listbox.Button className="flex w-[9em] flex-row justify-center self-center border-2 pl-2 pr-4 text-center">
           <InlineIcon i={<SelectedType.icon />} />
           {SelectedType.roomType}
         </Listbox.Button>
         <Listbox.Options className="absolute overflow-auto bg-black">
           {roomTypes.map((Item) => (
             <Listbox.Option
-              className="cursor-pointer bg-black p-[2px] hover:bg-teal-800"
+              className="flex cursor-pointer flex-row items-center bg-black p-[2px] hover:bg-teal-800"
               key={Item.roomType}
               value={Item}
             >
@@ -110,8 +110,10 @@ const CardElement = ({
           <div className="basis-[6em] p-2">ROOM TYPE</div>
           {/* ↓ の z-10 は, 展開されたリストがdisabled状態のボタンを重なった時にボタンの方が上にくる、という現象を防ぐためのもの */}
           {/* https://ics.media/entry/200609/ */}
-          <div className="z-10 shrink grow">
-            <RoomTypeListBox selected={roomType} setSelected={setRoomType} />
+          <div className="z-10 flex shrink grow items-center">
+            <div>
+              <RoomTypeListBox selected={roomType} setSelected={setRoomType} />
+            </div>
           </div>
         </div>
