+++ frontend/src/features/DM/DmCard.tsx
@@ -8,9 +8,10 @@ import * as TD from '@/typedef';
 
 type DmModalProps = {
   user?: TD.displayUser;
+  bordered?: boolean;
 };
 
-export const DmCard = ({ user }: DmModalProps) => {
+export const DmCard = ({ user, bordered }: DmModalProps) => {
   const [mySocket] = useAtom(chatSocketAtom);
   const [dmRooms] = useAtom(dataAtom.dmRoomsAtom);
   const [content, setContent] = useState('');
@@ -48,14 +49,18 @@ export const DmCard = ({ user }: DmModalProps) => {
       console.log(data);
       mySocket?.emit('ft_tell', data);
     }
+    setContent('');
   };
   const isSendable = !!user && content !== '';
 
+  const fieldClassName = `shrink grow ${
+    bordered ? ' border-2 border-solid border-white' : ''
+  }`;
   return (
     <>
-      <div className="flex flex-row">
+      <div className="flex flex-row items-center">
         <FTTextField
-          className="shrink grow"
+          className={fieldClassName}
           value={content}
           placeholder="発言内容"
           onChange={(e) => setContent(e.target.value)}
